Moved _BMF_ HandleGraph (over 4000 lines!) to separate file
